Split VideoDecoder
